+++ Skin/Retail/SharedXML/SharedUIPanelTemplates.lua
@@ -394,6 +394,9 @@ do --[[ SharedXML\SharedUIPanelTemplates.xml ]]
     end
     function Skin.InsetFrameTemplate(Frame)
         Frame.NineSlice.Center = Frame.Bg
+        if Frame.debug then
+            Frame.NineSlice.debug = Frame.debug
+        end
         Skin.NineSlicePanelTemplate(Frame.NineSlice)
     end
     function Skin.DialogBorderNoCenterTemplate(Frame)
@@ -457,6 +460,9 @@ do --[[ SharedXML\SharedUIPanelTemplates.xml ]]
 
         Frame.TopTileStreaks:SetTexture("")
 
+        if Frame.debug then
+            Frame.NineSlice.debug = Frame.debug
+        end
         Frame.NineSlice.Center = Frame.Bg
         Skin.NineSlicePanelTemplate(Frame.NineSlice)
     end
@@ -469,6 +475,9 @@ do --[[ SharedXML\SharedUIPanelTemplates.xml ]]
     function Skin.ButtonFrameTemplate(Frame)
         Skin.PortraitFrameTemplate(Frame)
         Skin.InsetFrameTemplate(Frame.Inset)
+    end
+    function Skin.ButtonFrameTemplateMinimizable(Frame)
+        Skin.ButtonFrameTemplate(Frame)
     end
 
     function Skin.MagicButtonTemplate(Button)