+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-﻿## [Unreleased] ##
+﻿## [9.0.1.6] - 2020-11-16 ##
 ### Added ###
 
   * [retail] Skinned Covenant Renown
@@ -244,6 +244,7 @@
   * Error when opening azerite respec UI
 
 [Unreleased]: https://github.com/Haleth/Aurora/compare/master...develop
+[9.0.1.6]: https://github.com/Haleth/Aurora/compare/9.0.1.5...9.0.1.6
 [9.0.1.5]: https://github.com/Haleth/Aurora/compare/9.0.1.4...9.0.1.5
 [9.0.1.4]: https://github.com/Haleth/Aurora/compare/9.0.1.3...9.0.1.4
 [9.0.1.3]: https://github.com/Haleth/Aurora/compare/9.0.1.2...9.0.1.3